+++ qcodes/utils/dataset/doNd.py
@@ -1,12 +1,7 @@
 from qcodes.dataset.dond.do_1d import do1d
 from qcodes.dataset.dond.do_2d import do2d
 from qcodes.dataset.dond.do_nd import (
-    AbstractSweep,
-    ArraySweep,
-    LinSweep,
-    LogSweep,
     ParameterGroup,
-    UnsafeThreadingException,
     _conditional_parameter_set,
     _create_measurements,
     _extract_paramters_by_type_and_group,
@@ -31,6 +26,7 @@ from qcodes.dataset.dond.do_nd_utils import (
     _register_parameters,
     _set_write_period,
 )
+from qcodes.dataset.dond.sweeps import AbstractSweep, ArraySweep, LinSweep, LogSweep
 from qcodes.dataset.plotting import plot_and_save_image as plot
 
 # todo enable warning once new api is in release
@@ -38,6 +34,10 @@ from qcodes.dataset.plotting import plot_and_save_image as plot
 #     "qcodes.utils.dataset.doNd module is deprecated. "
 #     "Please update to import from qcodes.dataset"
 # )
+
+
+class UnsafeThreadingException(Exception):
+    pass
 
 __all__ = [
     "AbstractSweep",
